backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
openai>=1.58.0
google-genai>=1.5.0
chromadb>=0.6.0
python-multipart>=0.0.12
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
psycopg2-binary>=2.9.9
aiofiles>=24.1.0
python-dotenv>=1.0.1
httpx>=0.28.0
pytest>=8.3.0
pytest-asyncio>=0.25.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0,<5.0
PyJWT>=2.8.0
